07_llm_attacks_text/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
rank-bm25==0.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
beautifulsoup4==4.12.3
langgraph==0.2.28
langchain-core==0.3.20
